requirements.txt
absl-py==0.4.0
astor==0.7.1
backcall==0.1.0
bleach==1.5.0
boto==2.49.0
boto3==1.9.10
botocore==1.12.10
bz2file==0.98
catboost==0.12.2
certifi==2018.8.24
chardet==3.0.4
cloudpickle==0.5.3
cognitive-face==1.4.2
cycler==0.10.0
cymem==1.31.2
cytoolz==0.9.0.1
dask==0.18.2
decorator==4.3.0
dill==0.2.8.2
docutils==0.14
easydict==1.8
en-core-web-sm==2.0.0
entrypoints==0.2.3
enum34==1.1.6
gast==0.2.0
gensim==3.6.0
grpcio==1.14.1
h3==3.4.1
h5py==2.8.0
html5lib==0.9999999
idna==2.7
imgaug==0.2.6
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
jedi==0.12.1
Jinja2==2.10
jmespath==0.9.3
joblib==0.13.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mistune==0.8.3
msgpack==0.5.6
msgpack-numpy==0.4.4.1
mtcnn==0.0.8
murmurhash==0.28.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.3
notebook==5.6.0
numpy==1.14.5
opencv-contrib-python==3.2.0.8
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
plac==0.9.6
preshed==1.0.1
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==0.5.2
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
regex==2017.4.5
requests==2.19.1
s3transfer==0.1.13
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
smart-open==1.7.1
spacy==2.0.12
SQLAlchemy==1.2.14
tensorboard==1.10.0
tensorboardX==1.4
tensorflow==1.4.1
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
thinc==6.10.3
toolz==0.9.0
torch==0.4.1
torchtext==0.2.3
torchvision==0.2.1
tornado==5.1
tqdm==4.26.0
traitlets==4.3.2
ujson==1.35
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.0
wrapt==1.10.11
xgboost==0.81
xlrd==1.1.0
